requirements_frozen.txt
asttokens==2.4.1
boto3==1.34.5
botocore==1.34.5
-e git+https://github.com/int-brain-lab/paper-brain-wide-map.git@d02f5ae678550854e8a9178ee434e0535cda9405#egg=brainwidemap
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorlog==6.8.0
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dask==2023.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
flake8==6.1.0
fonttools==4.47.0
fsspec==2023.12.2
globus-sdk==3.33.0.post0
graphviz==0.20.1
ibl-neuropixel==0.8.1
iblatlas==0.3.0
ibllib==2.26
iblutil==1.7.2
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.0
iniconfig==2.0.0
ipython==8.18.1
jedi==0.19.1
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
labcams==0.7.1
lazy_loader==0.3
llvmlite==0.41.1
locket==1.0.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mtscomp==1.0.2
natsort==8.4.0
networkx==3.2.1
npTDMS==1.7.1
nptyping==2.5.0
numba==0.58.1
numpy==1.26.2
ONE-api==2.5.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.4
parso==0.8.3
partd==1.4.1
pexpect==4.9.0
phylib==2.4.3
Pillow==10.1.0
PIMS==0.6.1
plotly==5.24.1
pluggy==1.3.0
prompt-toolkit==3.0.43
Psychofit==1.0.0.post0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
Pygments==2.17.2
PyJWT==2.8.0
pynrrd==1.0.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.3
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
s3transfer==0.9.0
scikit-image==0.22.0
scikit-learn==1.3.2
scikit-video==1.1.11
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
slicerator==1.1.0
slidingRP==1.0.0
sparse==0.14.0
stack-data==0.6.3
threadpoolctl==3.2.0
tifffile==2023.12.9
tomli==2.0.1
toolz==0.12.0
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.12
wfield==0.3.7
zipp==3.17.0
